pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "fl-web-service"
version = "0.0.0"
description = "Web service layer for the federated plant disease classification thesis project"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.30",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg[binary]>=3.2",
    "pydantic-settings>=2.2",
    "email-validator>=2.1",
    "bcrypt>=4.1",
    "pyjwt>=2.8",
    # Inference + FL training stack (orchestrator spawns `flwr run fl_app/`).
    # Versions pinned to match the experiment venv in the flwr-репо so trained
    # weights load byte-identically here. Install torch via the nightly cu128
    # index (see README); pyproject can't express the URL.
    "pillow>=10.0",
    "torch>=2.11",
    "torchvision>=0.26",
    "timm>=0.9",
    "datasets>=2.0",
    "scikit-learn>=1.8",
    "flwr[simulation]>=1.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "httpx>=0.27",
    "ruff>=0.5",
    "mypy>=1.10",
]

[tool.setuptools]
packages = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
